refactor: implement helpers
