+++ src/handlers/os.js
@@ -1,0 +1,19 @@
+import os from "os";
+
+export const osMethods = {
+  EOL: () => {
+    console.log(os.EOL);
+  },
+  cpus: () => {
+    console.log(os.cpus());
+  },
+  homedir: () => {
+    console.log(os.userInfo().homedir);
+  },
+  username: () => {
+    console.log(os.userInfo().username);
+  },
+  architecture: () => {
+    console.log(os.arch());
+  }
+};
