+++ src/handlers/zip.js
@@ -1,7 +1,8 @@
 import * as zlib from "zlib";
 import * as fsPromises from "fs/promises";
-import { join, parse } from "path";
 import * as fs from "fs";
+import { parse } from "path";
+import { pathResolver } from "../helpers/pathResolver.js";
 
 export const compress = async (
   currentPath,
@@ -9,25 +10,14 @@ export const compress = async (
   targetFilePath
 ) => {
   try {
-    if (!sourceFilePath || !targetFilePath) {
-      throw new Error("Invalid input");
-    }
-    let oldFilePath, newFilePath;
-    if (sourceFilePath.search(/^[a-zA-Z][:]/gm) === 0) {
-      oldFilePath = sourceFilePath;
-    } else {
-      oldFilePath = join(currentPath, sourceFilePath);
-    }
-    if (targetFilePath.search(/^[a-zA-Z][:]/gm) === 0) {
-      newFilePath = targetFilePath;
-    } else {
-      newFilePath = join(currentPath, targetFilePath);
-    }
+    const oldFilePath = pathResolver(currentPath, sourceFilePath);
+    const newFilePath = pathResolver(currentPath, targetFilePath);
 
     await fsPromises.access(oldFilePath);
-    const { dir, base } = parse(newFilePath);
 
+    const { dir, base } = parse(newFilePath);
     const targetDirListing = await fsPromises.readdir(dir);
+
     if (targetDirListing.includes(base)) {
       throw new Error("Operation failed. File already exists");
     }
@@ -55,20 +45,8 @@ export const decompress = async (
   targetFilePath
 ) => {
   try {
-    if (!sourceFilePath || !targetFilePath) {
-      throw new Error("Invalid input");
-    }
-    let oldFilePath, newFilePath;
-    if (sourceFilePath.search(/^[a-zA-Z][:]/gm) === 0) {
-      oldFilePath = sourceFilePath;
-    } else {
-      oldFilePath = join(currentPath, sourceFilePath);
-    }
-    if (targetFilePath.search(/^[a-zA-Z][:]/gm) === 0) {
-      newFilePath = targetFilePath;
-    } else {
-      newFilePath = join(currentPath, targetFilePath);
-    }
+    const oldFilePath = pathResolver(currentPath, sourceFilePath);
+    const newFilePath = pathResolver(currentPath, targetFilePath);
 
     await fsPromises.access(oldFilePath);
 
@@ -78,6 +56,7 @@ export const decompress = async (
     if (targetDirListing.includes(base)) {
       throw new Error("Operation failed. File already exists");
     }
+
     fs.createReadStream(oldFilePath)
       .pipe(zlib.createBrotliDecompress())
       .pipe(fs.createWriteStream(newFilePath))
