--- src/helpers/help.js
@@ -1,1 +1,0 @@
-
