+++ src/index.js
@@ -4,109 +4,18 @@ import { fileURLToPath } from "url";
 import { dirname, join } from "path";
 import { stdin as input, stdout as output } from "process";
 
-import CommandHandler from "./commandHandler.js";
+import FsCommandHandler from "./handlers/fsCommandHandler.js";
 import { parseArgs } from "./helpers/args.js";
-import { calculateHash } from "./helpers/calcHash.js";
-import { compress, decompress } from "./helpers/zip.js";
-
-const args = parseArgs();
-const username = args["username"];
-let currentPath = os.homedir();
+import { osMethods } from "./handlers/os.js";
+import { calculateHash } from "./handlers/calcHash.js";
+import { compress, decompress } from "./handlers/zip.js";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
-const osMethods = {
-  EOL: () => {
-    console.log(os.EOL);
-  },
-  cpus: () => {
-    console.log(os.cpus());
-  },
-  homedir: () => {
-    console.log(os.userInfo().homedir);
-  },
-  username: () => {
-    console.log(os.userInfo().username);
-  },
-  architecture: () => {
-    console.log(os.arch());
-  }
-};
-
-async function parseLine(line) {
-  const parsedLine = line.split(" ");
-  const [command, ...args] = parsedLine;
-
-  try {
-    switch (true) {
-      case command === "help":
-        await CommandHandler.cat(__dirname, join("..", "HELP.txt"));
-        return;
-      case command === "os":
-        if (args.length !== 1 || !args[0].startsWith("--")) {
-          throw new Error("Invalid input");
-        }
-        const arg = args[0].slice(2);
-        const func = osMethods[arg];
-        if (!func) {
-          throw new Error("Invalid input");
-        }
-        func();
-        return;
-      case command === "hash":
-        await calculateHash(currentPath, args[0]);
-        return;
-      case command === "ls":
-        await CommandHandler.ls(currentPath);
-        break;
-      case command === "up":
-        const upperPath = await CommandHandler.cd(currentPath, "..");
-        if (upperPath) {
-          currentPath = upperPath;
-        }
-        break;
-      case command === "cd":
-        const newPath = await CommandHandler.cd(currentPath, args[0]);
-        if (newPath) {
-          currentPath = newPath;
-        }
-        break;
-      case command === "cat":
-        const data = await CommandHandler.cat(currentPath, args[0]);
-        return data;
-      case command === "add":
-        await CommandHandler.add(currentPath, args[0]);
-        break;
-      case command === "rn":
-        await CommandHandler.rn(currentPath, args[0], args[1]);
-        break;
-      case command === "cp":
-        const cpData = await CommandHandler.cp(
-          currentPath,
-          args[0],
-          args[1]
-        );
-        return cpData;
-      case command === "mv":
-        await CommandHandler.mv(currentPath, args[0], args[1]);
-        break;
-      case command === "rm":
-        await CommandHandler.rm(currentPath, args[0]);
-        break;
-      case command === "compress":
-        await compress(currentPath, args[0], args[1]);
-        return;
-      case command === "decompress":
-        await decompress(currentPath, args[0], args[1]);
-        return;
-      default:
-        throw new Error("Invalid input");
-    }
-  } catch (e) {
-    console.log(e.message);
-  }
-}
+const args = parseArgs();
+const username = args["username"];
+let currentPath = os.homedir();
 
 const init = async () => {
   const rl = readline.createInterface({ input, output });
@@ -123,9 +32,7 @@ const init = async () => {
     if (readData === true) {
       console.log("Success!");
     }
-    /*    
-      await readData?.pipe(output);
-    */
+
     console.log(`You are currently in ${currentPath}`);
   });
 
@@ -137,3 +44,83 @@ const init = async () => {
 };
 
 await init();
+
+async function parseLine(line) {
+  const parsedLine = line.split(" ");
+  const [command, ...args] = parsedLine;
+
+  try {
+    switch (true) {
+      case command === "help":
+        await FsCommandHandler.cat(__dirname, join("..", "HELP.txt"));
+        return;
+      case command === "os":
+        if (args.length !== 1 || !args[0].startsWith("--")) {
+          throw new Error("Invalid input");
+        }
+        const arg = args[0].slice(2);
+        const func = osMethods[arg];
+        if (!func) {
+          throw new Error("Invalid input");
+        }
+        func();
+        return;
+      case command === "hash":
+        await calculateHash(currentPath, args[0]);
+        return;
+      case command === "ls":
+        await FsCommandHandler.ls(currentPath);
+        return;
+      case command === "up":
+        const upperPath = await FsCommandHandler.cd(
+          currentPath,
+          ".."
+        );
+        if (upperPath) {
+          currentPath = upperPath;
+        }
+        return;
+      case command === "cd":
+        const newPath = await FsCommandHandler.cd(
+          currentPath,
+          args[0]
+        );
+        if (newPath) {
+          currentPath = newPath;
+        }
+        return;
+      case command === "cat":
+        const data = await FsCommandHandler.cat(currentPath, args[0]);
+        return data;
+      case command === "add":
+        await FsCommandHandler.add(currentPath, args[0]);
+        return;
+      case command === "rn":
+        await FsCommandHandler.rn(currentPath, args[0], args[1]);
+        return;
+      case command === "cp":
+        const cpData = await FsCommandHandler.cp(
+          currentPath,
+          args[0],
+          args[1]
+        );
+        return cpData;
+      case command === "mv":
+        await FsCommandHandler.mv(currentPath, args[0], args[1]);
+        return;
+      case command === "rm":
+        await FsCommandHandler.rm(currentPath, args[0]);
+        return;
+      case command === "compress":
+        await compress(currentPath, args[0], args[1]);
+        return;
+      case command === "decompress":
+        await decompress(currentPath, args[0], args[1]);
+        return;
+      default:
+        throw new Error("Invalid input");
+    }
+  } catch (e) {
+    console.log(e.message);
+  }
+}
